+++ __snapshots__/flex-logs.test.js
@@ -6,13 +6,13 @@ exports['flex logs without query by specifying a profile and existent serviceId 
 [debug] Response: GET http://localhost:3234/v2/data-links/12378kdl2/logs 200
 Count: 5
 
-containerId   message                                                                threshold  timestamp               
-------------  ---------------------------------------------------------------------  ---------  ------------------------
-3864f1602739  *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...                null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Running /etc/rc.local...                                           null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Runit started as PID 11                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  *** Booting runit daemon...                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  chown: changing ownership of '/dev/xconsole': Operation not permitted  null       2017-08-30T08:06:49.595Z
+containerId   timestamp                 threshold  message                                                
+------------  ------------------------  ---------  -------------------------------------------------------
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/rc.local...                           
+3864f1602739  2017-08-30T08:06:49.595Z  null       *** Runit started as PID 11                            
+3864f1602739  2017-08-30T08:06:49.595Z  warn       *** Booting runit daemon...                            
+3864f1602739  2017-08-30T08:06:49.595Z  info       {"name":"test","num":10}                               
 
 
 
@@ -35,13 +35,13 @@ exports['flex logs without query by specifying a profile when valid project is s
 [debug] Response: GET http://localhost:3234/v2/data-links/12378kdl2/logs 200
 Count: 5
 
-containerId   message                                                                threshold  timestamp               
-------------  ---------------------------------------------------------------------  ---------  ------------------------
-3864f1602739  *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...                null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Running /etc/rc.local...                                           null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Runit started as PID 11                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  *** Booting runit daemon...                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  chown: changing ownership of '/dev/xconsole': Operation not permitted  null       2017-08-30T08:06:49.595Z
+containerId   timestamp                 threshold  message                                                
+------------  ------------------------  ---------  -------------------------------------------------------
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/rc.local...                           
+3864f1602739  2017-08-30T08:06:49.595Z  null       *** Runit started as PID 11                            
+3864f1602739  2017-08-30T08:06:49.595Z  warn       *** Booting runit daemon...                            
+3864f1602739  2017-08-30T08:06:49.595Z  info       {"name":"test","num":10}                               
 
 
 
@@ -54,13 +54,13 @@ exports['flex logs without query by specifying a profile when invalid project is
 [debug] Response: GET http://localhost:3234/v2/data-links/12378kdl2/logs 200
 Count: 5
 
-containerId   message                                                                threshold  timestamp               
-------------  ---------------------------------------------------------------------  ---------  ------------------------
-3864f1602739  *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...                null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Running /etc/rc.local...                                           null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Runit started as PID 11                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  *** Booting runit daemon...                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  chown: changing ownership of '/dev/xconsole': Operation not permitted  null       2017-08-30T08:06:49.595Z
+containerId   timestamp                 threshold  message                                                
+------------  ------------------------  ---------  -------------------------------------------------------
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/rc.local...                           
+3864f1602739  2017-08-30T08:06:49.595Z  null       *** Runit started as PID 11                            
+3864f1602739  2017-08-30T08:06:49.595Z  warn       *** Booting runit daemon...                            
+3864f1602739  2017-08-30T08:06:49.595Z  info       {"name":"test","num":10}                               
 
 
 
@@ -104,13 +104,13 @@ exports['flex logs without query by not specifying profile nor credentials when 
 [debug] Response: GET http://localhost:3234/v2/data-links/12378kdl2/logs 200
 Count: 5
 
-containerId   message                                                                threshold  timestamp               
-------------  ---------------------------------------------------------------------  ---------  ------------------------
-3864f1602739  *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...                null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Running /etc/rc.local...                                           null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Runit started as PID 11                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  *** Booting runit daemon...                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  chown: changing ownership of '/dev/xconsole': Operation not permitted  null       2017-08-30T08:06:49.595Z
+containerId   timestamp                 threshold  message                                                
+------------  ------------------------  ---------  -------------------------------------------------------
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/rc.local...                           
+3864f1602739  2017-08-30T08:06:49.595Z  null       *** Runit started as PID 11                            
+3864f1602739  2017-08-30T08:06:49.595Z  warn       *** Booting runit daemon...                            
+3864f1602739  2017-08-30T08:06:49.595Z  info       {"name":"test","num":10}                               
 
 
 
@@ -144,14 +144,17 @@ exports['flex logs without query by not specifying profile nor credentials when 
       "_id": "59a67219a86d05de4f97ea31",
       "containerId": "3864f1602739e924b257c7082f9e5c7bcd0b43d8986efe3acb4581f4f94dcd65",
       "message": "*** Booting runit daemon...",
-      "threshold": null,
+      "threshold": "warn",
       "timestamp": "2017-08-30T08:06:49.595Z"
     },
     {
       "_id": "59a672197f2a6ba14ec962e8",
       "containerId": "3864f1602739e924b257c7082f9e5c7bcd0b43d8986efe3acb4581f4f94dcd65",
-      "message": "chown: changing ownership of '/dev/xconsole': Operation not permitted",
-      "threshold": null,
+      "message": {
+        "name": "test",
+        "num": 10
+      },
+      "threshold": "info",
       "timestamp": "2017-08-30T08:06:49.595Z"
     }
   ]
@@ -201,13 +204,13 @@ exports['flex logs without query by specifying credentials as options when valid
 [debug] Logged out current user.
 Count: 5
 
-containerId   message                                                                threshold  timestamp               
-------------  ---------------------------------------------------------------------  ---------  ------------------------
-3864f1602739  *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...                null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Running /etc/rc.local...                                           null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Runit started as PID 11                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  *** Booting runit daemon...                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  chown: changing ownership of '/dev/xconsole': Operation not permitted  null       2017-08-30T08:06:49.595Z
+containerId   timestamp                 threshold  message                                                
+------------  ------------------------  ---------  -------------------------------------------------------
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/rc.local...                           
+3864f1602739  2017-08-30T08:06:49.595Z  null       *** Runit started as PID 11                            
+3864f1602739  2017-08-30T08:06:49.595Z  warn       *** Booting runit daemon...                            
+3864f1602739  2017-08-30T08:06:49.595Z  info       {"name":"test","num":10}                               
 
 
 
@@ -244,13 +247,13 @@ exports['flex logs with query with valid timestamps and valid paging should succ
 [debug] Response: GET http://localhost:3234/v2/data-links/12378kdl2/logs?from=2017-08-30T08:06:49.594Z&to=2017-09-02T08:06:49&limit=5&page=3 200
 Count: 5
 
-containerId   message                                                                threshold  timestamp               
-------------  ---------------------------------------------------------------------  ---------  ------------------------
-3864f1602739  *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...                null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Running /etc/rc.local...                                           null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Runit started as PID 11                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  *** Booting runit daemon...                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  chown: changing ownership of '/dev/xconsole': Operation not permitted  null       2017-08-30T08:06:49.595Z
+containerId   timestamp                 threshold  message                                                
+------------  ------------------------  ---------  -------------------------------------------------------
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/rc.local...                           
+3864f1602739  2017-08-30T08:06:49.595Z  null       *** Runit started as PID 11                            
+3864f1602739  2017-08-30T08:06:49.595Z  warn       *** Booting runit daemon...                            
+3864f1602739  2017-08-30T08:06:49.595Z  info       {"name":"test","num":10}                               
 
 
 
@@ -264,13 +267,13 @@ exports['flex logs with query with valid timestamps and without paging should su
 [debug] Response: GET http://localhost:3234/v2/data-links/12378kdl2/logs?from=2017-08-30T08:06:49.594Z&to=2017-09-02T08:06:49.000Z 200
 Count: 5
 
-containerId   message                                                                threshold  timestamp               
-------------  ---------------------------------------------------------------------  ---------  ------------------------
-3864f1602739  *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...                null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Running /etc/rc.local...                                           null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Runit started as PID 11                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  *** Booting runit daemon...                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  chown: changing ownership of '/dev/xconsole': Operation not permitted  null       2017-08-30T08:06:49.595Z
+containerId   timestamp                 threshold  message                                                
+------------  ------------------------  ---------  -------------------------------------------------------
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/rc.local...                           
+3864f1602739  2017-08-30T08:06:49.595Z  null       *** Runit started as PID 11                            
+3864f1602739  2017-08-30T08:06:49.595Z  warn       *** Booting runit daemon...                            
+3864f1602739  2017-08-30T08:06:49.595Z  info       {"name":"test","num":10}                               
 
 
 
@@ -292,13 +295,13 @@ exports['flex logs with query with valid start timestamp and nothing else should
 [debug] Response: GET http://localhost:3234/v2/data-links/12378kdl2/logs?from=2017-08-30T08:06:49.594Z 200
 Count: 5
 
-containerId   message                                                                threshold  timestamp               
-------------  ---------------------------------------------------------------------  ---------  ------------------------
-3864f1602739  *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...                null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Running /etc/rc.local...                                           null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Runit started as PID 11                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  *** Booting runit daemon...                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  chown: changing ownership of '/dev/xconsole': Operation not permitted  null       2017-08-30T08:06:49.595Z
+containerId   timestamp                 threshold  message                                                
+------------  ------------------------  ---------  -------------------------------------------------------
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/rc.local...                           
+3864f1602739  2017-08-30T08:06:49.595Z  null       *** Runit started as PID 11                            
+3864f1602739  2017-08-30T08:06:49.595Z  warn       *** Booting runit daemon...                            
+3864f1602739  2017-08-30T08:06:49.595Z  info       {"name":"test","num":10}                               
 
 
 
@@ -328,13 +331,13 @@ exports['flex logs with query without timestamps and valid paging should succeed
 [debug] Response: GET http://localhost:3234/v2/data-links/12378kdl2/logs?limit=5&page=3 200
 Count: 5
 
-containerId   message                                                                threshold  timestamp               
-------------  ---------------------------------------------------------------------  ---------  ------------------------
-3864f1602739  *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...                null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Running /etc/rc.local...                                           null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Runit started as PID 11                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  *** Booting runit daemon...                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  chown: changing ownership of '/dev/xconsole': Operation not permitted  null       2017-08-30T08:06:49.595Z
+containerId   timestamp                 threshold  message                                                
+------------  ------------------------  ---------  -------------------------------------------------------
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/rc.local...                           
+3864f1602739  2017-08-30T08:06:49.595Z  null       *** Runit started as PID 11                            
+3864f1602739  2017-08-30T08:06:49.595Z  warn       *** Booting runit daemon...                            
+3864f1602739  2017-08-30T08:06:49.595Z  info       {"name":"test","num":10}                               
 
 
 
@@ -348,13 +351,13 @@ exports['flex logs with query without timestamps and page but with valid size sh
 [debug] Response: GET http://localhost:3234/v2/data-links/12378kdl2/logs?limit=35 200
 Count: 5
 
-containerId   message                                                                threshold  timestamp               
-------------  ---------------------------------------------------------------------  ---------  ------------------------
-3864f1602739  *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...                null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Running /etc/rc.local...                                           null       2017-08-30T08:06:49.594Z
-3864f1602739  *** Runit started as PID 11                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  *** Booting runit daemon...                                            null       2017-08-30T08:06:49.595Z
-3864f1602739  chown: changing ownership of '/dev/xconsole': Operation not permitted  null       2017-08-30T08:06:49.595Z
+containerId   timestamp                 threshold  message                                                
+------------  ------------------------  ---------  -------------------------------------------------------
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/my_init.d/00_regen_ssh_host_keys.sh...
+3864f1602739  2017-08-30T08:06:49.594Z  null       *** Running /etc/rc.local...                           
+3864f1602739  2017-08-30T08:06:49.595Z  null       *** Runit started as PID 11                            
+3864f1602739  2017-08-30T08:06:49.595Z  warn       *** Booting runit daemon...                            
+3864f1602739  2017-08-30T08:06:49.595Z  info       {"name":"test","num":10}